requirements.txt
anyio==4.3.0
black==24.2.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pymongo==4.6.2
python-dotenv==1.0.1
python-telegram-bot==20.8
sniffio==1.3.1
tomli==2.0.1
typing_extensions==4.10.0
